+++ game/hooks/chargen/custom_app_review.rb
@@ -1,0 +1,11 @@
+module AresMUSH
+  module Chargen
+    def self.custom_app_review(char)
+      if Manage.is_extra_installed?('osr_rpg')
+        return OsrRpg.app_review(char)
+      end
+
+      nil
+    end
+  end
+end